testing

+++ testing.py
@@ -364,7 +364,7 @@ for l in range(layers):
 model = load_model('../models_3/model_ref_back.h5')
 pred_back = np.squeeze(yhat_back.detach().numpy())
 print(pred_back.shape)
-img = Image.fromarray(pred_back)
+img = Image.fromarray(pred_back.astype(np.uint8))
 # pixels = img_to_array(pixels)
 # pixels = (pixels - 127.5) / 127.5
 # img = expand_dims(pixels, 0)
